query builder: re-organize projection state
